requirements.txt
fastapi
uvicorn
yfinance
requests
yahooquery
python-dotenv
google-auth
google-auth-oauthlib
python-jose[cryptography]
python-multipart
pydantic-settings
aiohttp